warn_once to logger_core
