+++ mephisto/data_model/assignment.py
@@ -23,10 +23,9 @@ if TYPE_CHECKING:
 
 import os
 import json
-from mephisto.utils.misc import warn_once
 from dataclasses import dataclass
 
-from mephisto.utils.logger_core import get_logger
+from mephisto.utils.logger_core import get_logger, warn_once
 
 logger = get_logger(name=__name__)
