+++ mephisto/data_model/worker.py
@@ -12,8 +12,7 @@ from mephisto.data_model.db_backed_meta import (
     MephistoDataModelComponentMixin,
 )
 from typing import Any, List, Optional, Mapping, Tuple, Dict, Type, Tuple, TYPE_CHECKING
-from mephisto.utils.logger_core import get_logger
-from mephisto.utils.misc import warn_once
+from mephisto.utils.logger_core import get_logger, warn_once
 
 logger = get_logger(name=__name__)
