+++ mephisto/utils/misc.py
@@ -6,39 +6,10 @@
 
 import logging
 import os
-from typing import Set, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from mephisto.data_model.requester import Requester
-
-
-_seen_logs: Set[str] = set()
-
-
-def warn_once(msg: str) -> None:
-    """
-    Log a warning, but only once.
-
-    :param str msg: Message to display
-    """
-    global _seen_logs
-    if msg not in _seen_logs:
-        _seen_logs.add(msg)
-        logging.warn(msg)
-
-
-def ensure_user_confirm(display_text, skip_input=False) -> None:
-    """
-    Helper to provide the flow for having a user confirm a specific occurrence
-    before it happens. skip_input will make this method return without
-    checking, which is useful for automated scripts
-    """
-    if skip_input:
-        return
-    res = input(f'{display_text}\nEnter "n" to exit and anything else to continue:')
-    if res == "n":
-        raise SystemExit(0)
-    return
 
 
 def get_mock_requester(db) -> "Requester":
